+++ routes/web.php
@@ -14,3 +14,9 @@
 Route::get('/', function () {
     return view('home');
 });
+Route::get('loginOAuth', 'LoginController@loginOAuth')->middleware('guest');
+Route::get('logout', 'LoginController@logout')->middleware('auth');
+Route::get('login', function () {
+    return view('login');
+})->middleware('guest');
+Route::post('login', 'LoginController@login')->middleware('guest');