+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Every changes made to the project will be documented here.
 
+## [0.3.0] - 31-05-2019
+## Added
+ - OAuth2 login with a-capella.ch
+ - Login with simple Auth
+ - Error handling
+
+## Changed
+ - Fixed issue in reactivity with carousel
+ - Colours in the site
+ 
 ## [0.2.1] - 23-05-2019
 ## Added
  - Phonetics were added in the cards